md/raid1: avoid overflow in raid1 resync when bitmap is in use.
Patchwork-id: 31759
O-Subject: [RHEL6.1 Patch 61/80] md/raid1: avoid overflow in raid1 resync when
	bitmap is in use.
Bugzilla: 633695 659623
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

bitmap_start_sync returns - via a pass-by-reference variable - the
number of sectors before we need to check with the bitmap again.
Since commit ef4256733506f245 this number can be substantially larger,
2^27 is a common value.

Unfortunately it is an 'int' and so when raid1.c:sync_request shifts
it 9 places to the left it becomes 0.  This results in a zero-length
read which the scsi layer justifiably complains about.

This patch just removes the shift so the common case becomes safe with
a trivially-correct patch.

In the next merge window we will convert this 'int' to a 'sector_t'

Reported-by: "George Spelvin" <linux@horizon.com>
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>