kabi: update check-kabi script to catch removed symbols
Patchwork-id: 31812
O-Subject: [RHEL6.1 PATCH 1/2] kabi: update check-kabi script to catch removed
	symbols
Bugzilla: 639815
RH-Acked-by: Don Zickus <dzickus@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=639815

The existing script used to check for kABI breaks during a kernel build
did not handle the removal of symbols. In RHEL5, this was detected in
RHTS. In RHEL6, since all symbol data is available at the RPM level,
it is flagged up in rpmdiff. Still, that is error-prone and way too
late to catch such problems. Therefore, add this check during build.

Signed-off-by: Jon Masters <jcm@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>