qdio: prevent race for shared indicators
Patchwork-id: 31818
O-Subject: [RHEL6.1 PATCH 4/6] [s390x] qdio: prevent race for shared indicators
Bugzilla: 668835

Description
-----------
Traffic on qdio devices might hang forever.

The shared DSCI is set at the end of the interrupt handler
to scan the queues that use the shared DSCI for further
traffic.  Because interrupts might be avoided if the DSCI
is set this can lead to stalled queues.  Also the usage of
the ALSI to detect further incoming traffic on the shared
DSCI queues is not race safe.

Only reset the shared DSCI if it was set at the beginning
of the interrupt handler so the tasklets for all queues that
use the shared DSCI are scheduled to run after the interrupt
handler.

Bugzilla
--------
BZ 668835
https://bugzilla.redhat.com/show_bug.cgi?id=668835

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.38
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=4f325184f2d4c1f2258873b2a333005dc4dfcbc0

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>