qeth: wait for recovery finish in open function
Patchwork-id: 31820
O-Subject: [RHEL6.1 PATCH 5/6] [s390x] qeth: wait for recovery finish in open
	function
Bugzilla: 668845

Description
-----------
Device initialization of a qeth device contains a STARTLAN step.
It might fail, but initialization continues and if configured even
the network interface is opened and set up.  Once cable is plugged
in, a qeth device recovery is triggered with successful STARTLAN.
A recovery switches from state SOFTSETUP to HARDSETUP to DOWN to
HARDSETUP and back to SOFTSETUP. If open and recover are running
concurrently, open fails if it hits the states HARDSETUP or DOWN.

Add waiting for recovery finish in the qeth open function to enable
successful qeth device opening in spite of a running recovery.

Bugzilla
--------
BZ 668845
https://bugzilla.redhat.com/show_bug.cgi?id=668845

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.38
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=2b6203bb7d85e6a2ca2088b8684f30be70246ddf

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>