tg3: Disable TSS except for 5719
Patchwork-id: 31949
O-Subject: [RHEL6.1 PATCH] tg3: Disable TSS except for 5719
Bugzilla: 634316
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

bz634316
https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=634316
tg3: Disable TSS

Description of problem:
TSS has been found by Broadcom to lock up the system with specific
tg3 devices. The only current workaround is to disable it, unless
the device is a 5719.

Solution:
Two patches are provided, one disable TSS for all devices and the
other patch qualifies the disabling so as to allow TSS on a 5719.

The upstream commits are:
  1. Disable TSS
     commit: f0392d24485e4c144d2741a5424f069b8b5f30e7
  2. Reenable TSS for 5719
     commit: b92b9040f6e4997b895b7b9c655a158354d28964

Note: The patch provided was built on top of the patch set to update
tg3 (bz635078) because that is the way RHEL6.1 should be shipping.

Brew: Successfully built in Brew for all architectures (task_3020022).

Testing:
I successfully ran my sanity tests. Waiting for test results from Broadcom.
When these results become available, I will provide an update.

Acks would be appreciated. Thanks.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>