cxgb3i: fixed connection problem with iscsi private ip
Patchwork-id: 31959
O-Subject: [RHEL 6.1 PATCH 4/7] cxgb3i: fixed connection problem with iscsi
	private ip
Bugzilla: 668632
RH-Acked-by: Rob Evers <revers@redhat.com>

From: kxie@chelsio.com <kxie@chelsio.com>

This is for bz 668632.
It is not merged upstream yet. It has been sent here:
http://groups.google.com/group/open-iscsi/browse_thread/thread/2162ab93534adedb

From: Karen Xie <kxie@chelsio.com>

fixed the connection problem when the private iscsi ipv4 address is provisioned on the interface. Without this patch we cannot connect to the target when
using a different IP for the iscsi interface than is used for the network
one.

Signed-off-by: Karen Xie <kxie@chelsio.com>
Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>