cxgbi: enable TEXT PDU support
Patchwork-id: 31958
O-Subject: [RHEL 6.1 PATCH 5/7] cxgbi: enable TEXT PDU support
Bugzilla: 668632
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 668632.

It has been sent upstream for the next feature window
[no link because googlegroups topic list is not updating].

cxgb3i and cxgb4i support TEXT PDU offloading, so set
the bits to enable it so we can do discovery through
the card instead of having to setup a seperate nic
for it.

Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>