xfs: serialise unaligned direct IOs
Patchwork-id: 31968
O-Subject: [RHEL6.1 PATCH 2/3] xfs: serialise unaligned direct IOs
Bugzilla: 669272
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>

RH BZ: 669272
Upstream Commit: pending. Review complete.

This is a backport of the last patch in the upstream unaligned
aio/dio corruption fix series. The cleanup patches do not apply, so
I've ported this bug fix patch alone to the rhel6 tree. The code is
a bit different as a result, but the same logic steps are taken to
fix the bug. Passes xfstests.

----
When two concurrent unaligned, non-overlapping direct IOs are issued
to the same block, the direct Io layer will race to zero the block.
The result is that one of the concurrent IOs will overwrite data
written by the other IO with zeros. This is demonstrated by the
xfsqa test 240.

To avoid this problem, serialise all unaligned direct IOs to an
inode with a big hammer. We need a big hammer approach as we need to
serialise AIO as well, so we can't just block writes on locks.
Hence, the big hammer is calling xfs_ioend_wait() while holding out
other unaligned direct IOs from starting.

We don't bother trying to serialised aligned vs unaligned IOs as
they are overlapping IO and the result of concurrent overlapping IOs
is undefined - the result of either IO is a valid result so we let
them race. Hence we only penalise unaligned IO, which already has a
major overhead compared to aligned IO so this isn't a major problem.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Alex Elder <aelder@sgi.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>