cifs: add cruid= mount option
Patchwork-id: 31974
O-Subject: [RHEL6 PATCH] BZ#667654: cifs: add cruid= mount option
Bugzilla: 667654
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

This patch fixes a subtle bug that crops up if you try to make
multiple krb5 mounts to the same server using different credentials.
The patch was just sent the other day, but Steve French has indicated
that he's planning to push this for 2.6.38.

It fixes the problem in my testing. The reporter of the problem is in
the process of testing it, but I haven't heard results back yet. Still
it's pretty straightforward and I'll be quite surprised if it doesn't
fix the problem.

-----------------------[snip]-------------------------

In commit 3e4b3e1f we separated the "uid" mount option such that it
no longer determined the owner of the credential cache by default. When
we did this, we added a new option to cifs.upcall (--legacy-uid) to
try to make it so that it would behave the same was as it did before.

This ignored a rather important point -- the kernel has no way to know
what options are being passed to cifs.upcall, so it doesn't know what
uid it should use to determine whether to match an existing krb5 session.

The simplest solution is to simply add a new "cruid=" mount option that
only governs the uid owner of the credential cache for the mount.

Unfortunately, this means that the --legacy-uid option in cifs.upcall was
ill-considered and is now useless, but I don't see a better way to deal
with this.

A patch for the mount.cifs manpage will follow once this patch has been
accepted.

Reported-by: Stefan Walter <walteste@inf.ethz.ch>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Cc: stable@kernel.org
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>