mpt2sas: Copy message frame before releasing to free pool to have a local reference
Patchwork-id: 32044
O-Subject: [RHEL6.1 PATCH 15/34] Copy message frame before releasing to free
	pool to have a local reference
Bugzilla: 642590
RH-Acked-by: Rob Evers <revers@redhat.com>

Current driver is  not clearing the per device tm_busy flag
following the Task Mangement request completion from the IOCTL path.
When this flag is set, the IO queues are frozen.   The reason the flag
didn't get cleared is becuase the driver is referencing
memory associated to the mpi request following the completion, when
the memory had been reallocated for a new request.  When the memory
was reallocated, the driver didn't clear the flag becuase it was
expecting a task managment reqeust, and the reallocated request was
for SCSI_IO.  To fix the problem the driver needs to have a cached
backup copy of the original reqeust.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>