mpt2sas: Device removal handshake even though the PHYSTATUS_VACANT bit is set in the PhyStatus
Patchwork-id: 32047
O-Subject: [RHEL6.1 PATCH 24/34] Device removal handshake even though the
	PHYSTATUS_VACANT bit is set in the PhyStatus
Bugzilla: 642590
RH-Acked-by: Rob Evers <revers@redhat.com>

When zoning end devices, the driver is not sending device
removal handshake alogrithm to firmware. This results in controller
firmware not sending sas topology add events the next time the device is
added. The fix is the driver should be doing the device removal handshake
even though the PHYSTATUS_VACANT bit is set in the PhyStatus of the
event data. The current design is avoiding the handshake when the
VACANT bit is set in the phy status.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>