mpt2sas: Call the_scsih_ir_shutdown prior to reporting the volumes missing from the OS
Patchwork-id: 32045
O-Subject: [RHEL6.1 PATCH 32/34] Call the_scsih_ir_shutdown prior to reporting
	the volumes missing from the OS
Bugzilla: 642590
RH-Acked-by: Rob Evers <revers@redhat.com>

Issue:
IR shutdown(sending) and IR shutdown(complete) messages not
listed in /var/log/messages when driver is removed.

The driver needs to issue a MPI2_RAID_ACTION_SYSTEM_SHUTDOWN_INITIATED
request when the driver is unloaded so the IR metadata journal is updated.
If this request is not sent, then the volume would need a "check
consistency" issued on the next bootup if the volume was roamed from one
initiator to another. The current driver supports this feature only when the
system is rebooted, however this also need to be supported if the driver is
unloaded

Fix:
To fix this issue, the driver is going
to need to call the _scsih_ir_shutdown prior to reporting
the volumes missing from the OS, hence the device handles
are still present.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>