Explicitly rule out llseek to avoid BKL in default_llseek()
Patchwork-id: 32079
O-Subject: [RHEL6.1 Patch 019/386] IB: Explicitly rule out llseek to avoid BKL
	in default_llseek()
Bugzilla: 607396 633157 633329 633875 635914 660674 660680
CVE: CVE-2010-3296
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Roland Dreier <rolandd@cisco.com>

Several RDMA user-access drivers have file_operations structures with
no .llseek method set.  None of the drivers actually do anything with
f_pos, so this means llseek is essentially a NOP, instead of returning
an error as leaving other file_operations methods unimplemented would
do.  This is mostly harmless, except that a NULL .llseek means that
default_llseek() is used, and this function grabs the BKL, which we
would like to avoid.

Since llseek does nothing useful on these files, we would like it to
return an error to userspace instead of silently grabbing the BKL and
succeeding.  For nearly all of the file types, we take the
belt-and-suspenders approach of setting the .llseek method to
no_llseek and also calling nonseekable_open(); the exception is the
uverbs_event files, which are created with anon_inode_getfile(), which
already sets f_mode the same way as nonseekable_open() would.

This work is motivated by Arnd Bergmann's bkl-removal tree.

Signed-off-by: Roland Dreier <rolandd@cisco.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>