rds: Fix RDMA message reference counting
Patchwork-id: 32426
O-Subject: [RHEL6.1 Patch 241/386] rds: Fix RDMA message reference counting
Bugzilla: 607396 633157 633329 633875 635914 660674 660680
CVE: CVE-2010-3296
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Chris Mason <chris.mason@oracle.com>

The RDS send_xmit code was trying to get fancy with message
counting and was dropping the final reference on the RDMA messages
too early.  This resulted in memory corruption and oopsen.

The fix here is to always add a ref as the parts of the message passes
through rds_send_xmit, and always drop a ref as the parts of the message
go through completion handling.

Signed-off-by: Chris Mason <chris.mason@oracle.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>