RDS: Let rds_message_alloc_sgs() return NULL
Patchwork-id: 32425
O-Subject: [RHEL6.1 Patch 369/386] RDS: Let rds_message_alloc_sgs() return NULL
Bugzilla: 607396 633157 633329 633875 635914 660674 660680
CVE: CVE-2010-3296
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Andy Grover <andy.grover@oracle.com>

Even with the previous fix, we still are reading the iovecs once
to determine SGs needed, and then again later on. Preallocating
space for sg lists as part of rds_message seemed like a good idea
but it might be better to not do this. While working to redo that
code, this patch attempts to protect against userspace rewriting
the rds_iovec array between the first and second accesses.

The consequences of this would be either a too-small or too-large
sg list array. Too large is not an issue. This patch changes all
callers of message_alloc_sgs to handle running out of preallocated
sgs, and fail gracefully.

Signed-off-by: Andy Grover <andy.grover@oracle.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>