mlx4: Fix IBoE reported link rate
Patchwork-id: 32431
O-Subject: [RHEL6.1 Patch 378/386] IB/mlx4: Fix IBoE reported link rate
Bugzilla: 607396 633157 633329 633875 635914 660674 660680
CVE: CVE-2010-3296
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Eli Cohen <eli@dev.mellanox.co.il>

The link rate is the product of the link speed in the link width. For
Etherent ports the rate is 10G, so we use 1 for the width and 4 for
speed to get the correct rate.

Signed-off-by: Eli Cohen <eli@mellanox.co.il>
Signed-off-by: Roland Dreier <rolandd@cisco.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>