bfa rebase reflecting scsi-misc bfa
Patchwork-id: 32467
O-Subject: [RHEL6.1 PATCH] bfa rebase reflecting scsi-misc bfa
Bugzilla: 641052
RH-Acked-by: Mike Christie <mchristi@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=641052

This patch is a rebase of bfa driver reflecting significant restructuring
and cleanup of the driver that has occurred upstream over the past
several months.  Several bug fixes are included.

This patch brings the rhel6 version of the bfa driver almost in sync with the
upstream version currently in scsi misc up to and including:

8f4bfadd200477acb6dcf41a45919dd37d01a6db.

I will be providing Brocade with some feedback regarding some nits for
future updates.

The bfa driver is currently tech preview for all rhel releases.  The tech
preview status is constantly being evaluated.  Any review of the driver
with this patch would be greatly appreciated for future work by Brocade on
the bfa driver.

A /lib/firmware update is occurring simultaneously to this patch.  Both
depend on each other.  The firmware bz is 617017.

The bna networking driver is slated to be included in rhel6 and also depends
on the firmware bz 617017 as it shares the firmware image with the bfa when
running on a Brocade CNA.

Testing was done at Brocade and sanity testing is under way using dt at RH.

http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3033081
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>