dvb/rc: update to dvb code from 2.6.38-rc1
Patchwork-id: 32544
O-Subject: [RHEL6 PATCH v2 02/14] v4l/dvb/rc: update to dvb code from 2.6.38-rc1
Bugzilla: 663280
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Mauro Carvalho Chehab <mchehab@redhat.com>

Bug 663280 - Update drivers/media to 2.6.38 codebase

This is a drop-in of all the current 2.6.38-rc1-to-be v4l/dvb code
specific to the dvb portions of the subsystem, pulled from the
linuxtv.org media_tree.git staging/for_2.6.38-rc1 branch.

Note: as of this changeset, the tree will NOT compile, as there are
forthcoming backports patches to cope with the delta between our 2.6.32
codebase and the 2.6.38-rc1 codebase this code originates from.

Signed-off-by: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>