nmi: Add in new nmi_watchdog/softlockup changes
Patchwork-id: 32574
O-Subject: [RHEL6 PATCH 2/6] x86, nmi: Add in new nmi_watchdog/softlockup changes
Bugzilla: 669808
RH-Acked-by: Peter Zijlstra <pzijlstr@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=669808

This change updates to the new generic nmi_watchdog (now called
hardlockup) and merges in the softlockup code.

The old x86 specific nmi_watchdog hard coded performance counter
allocation and its place on the bottom of a notifier chain.  This
is starting to cause issues for tools like perf and other nmi drivers
by vendors that want to be on the bottom of the NMI chain collecting
NMIs.

The new nmi_watchdog has moved to a new generic location and registers
with the perf subsystem.  This allows perf to manage the performance
counter allocation (the watchdog can even share counters, thanks to
the perf scheduler).  The imlementation is alot cleaner and with the
merge with the softlockup code, alot smaller.

The change is based off of 2.6.37.  It has been sitting upstream (and
running in Fedora) since 2.6.36, so it has been well tested.  The bulk
of the change involves adding kernel/watchdog.c and arch/x86/kernel/apic/
hw_nmi.c, while removing kernel/softlockup.c.  The rest of the changes
involve properly hooking into the Kconfig system and renaming CONFIGs
to LOCKUP_DETECTOR.

This patch is dependent on the recently posted perf update patch.

I would post a brew build id but, it took multiple attempts due to
brew overload to make sure it compiled fine on s390 and any new attempts
to build probably won't be done until next.  Once I get one I will post
it though.

Testing on various machines to make sure the box crashes on a hardlockup.

Note: This still preserves the old nmi watchdog.  In case the new has
serious problems or causes headaches, we can do a CONFIG change to revert
to the old behaviour.

Note: It has come to my recent attention that some vendors have programmed
their BIOSes to use performance counters.  With the new perf subsystem
update, this will block the nmi watchdog from running as perf does not
have the ability to share resources with the BIOS.  Then again, the
alternative is to just piggy back on those perf counters and scratch
our head when weird stuff happens (as we have been doing in RHEL-4/5/6).
Therefore the new crop of warnings that will pop up as a result of this
change is something that has been a problem for years, we just never
understood why.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>